Renamed app.js to index and app.css to style
